+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Added**
 - Commenting support for PostgreSQL.
   - With "has:comment" query string support.
+  - And "comment:SOME_STRING" for search comments.
 
 [Full Changelog](https://github.com/jasonish/evebox/compare/0.8.0...HEAD)